+++ fiborule/fiborule-common/src/main/java/com/fibo/rule/common/model/ChannelInfo.java
@@ -4,11 +4,11 @@ import lombok.Data;
 
 @Data
 public final class ChannelInfo {
-    private int app;
+    private Long app;
     private String address;
     private long lastUpdateTime;
 
-    public ChannelInfo(int app, String address, long lastUpdateTime) {
+    public ChannelInfo(Long app, String address, long lastUpdateTime) {
         this.app = app;
         this.address = address;
         this.lastUpdateTime = lastUpdateTime;
